requirements.txt
asgiref==3.11.1
Django==4.2.29
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
pyotp==2.9.0
sqlparse==0.5.5
typing_extensions==4.15.0
